requirements.txt
solana
requests
python-telegram-bot
openai
gspread
oauth2client